// src/app/models/report.ts
import { AssociateSkill } from "./associateSkill";

export class Report {
    skillSpread: AssociateSkill[];
    registeredCandidate: number;
    femaleCandidatePercentage: number;
    maleCandidatePercentage: number;
    fresherCandidatePercentage: number;
    ratedCandidate: number;
    femaleRatedCandidatePercentage: number;
    maleRatedCandidatePercentage: number;
    l1CandidatePercentage: number;
    l2CandidatePercentage: number;
    l3CandidatePercentage: number;
}